+++ src/project/schemas/recipe_product.py
@@ -4,7 +4,5 @@ from pydantic import BaseModel, ConfigDict
 class ProductSchema(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
-    id: int
-    name: str
-    cost: int
-
+    id_recipe: id
+    id_product: id